# backend/services/account-service/api/docs/swagger.yaml
...
openapi: 3.0.0
info:
  title: Account Service API
  description: API для управления аккаунтами, профилями и настройками пользователей
  version: 1.0.0
  contact:
    name: Gaiming Team
    url: https://gaiming.ru
    email: support@gaiming.ru
servers:
  - url: https://api.gaiming.ru/v1
    description: Production server
  - url: https://staging-api.gaiming.ru/v1
    description: Staging server
  - url: http://localhost:8080/v1
    description: Local development server
...
paths:
...
  /profiles:
    get:
      summary: Получение списка профилей
      description: Возвращает список профилей с пагинацией и фильтрацией
      tags:
        - profiles
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: visibility
          in: query
          description: Фильтр по видимости профиля
          schema:
            type: string
            enum: [public, friends, private]
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
  /profiles/{id}:
    get:
      summary: Получение профиля по ID
      description: Возвращает информацию о профиле по его ID
      tags:
        - profiles
      parameters:
        - $ref: '#/components/parameters/ProfileIdParam'
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
    
    put:
      summary: Обновление профиля
      description: Обновляет информацию о профиле
      tags:
        - profiles
      parameters:
        - $ref: '#/components/parameters/ProfileIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileUpdate'
      responses:
        '200':
          description: Профиль успешно обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
  /profiles/{id}/history:
    get:
      summary: Получение истории изменений профиля
      description: Возвращает историю изменений профиля
      tags:
        - profiles
      parameters:
        - $ref: '#/components/parameters/ProfileIdParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: field_name
          in: query
          description: Фильтр по имени поля
          schema:
            type: string
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileHistoryList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
...
components:
  schemas:
...
    Profile:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Уникальный идентификатор профиля
        account_id:
          type: string
          format: uuid
          description: Идентификатор аккаунта
        nickname:
          type: string
          description: Отображаемое имя пользователя
        bio:
          type: string
          description: Биография пользователя
        country:
          type: string
          description: Страна пользователя
        city:
          type: string
          description: Город пользователя
        birth_date:
          type: string
          format: date
          description: Дата рождения пользователя
        gender:
          type: string
          description: Пол пользователя
        visibility:
          type: string
          enum: [public, friends, private]
          description: Видимость профиля
        created_at:
          type: string
          format: date-time
          description: Дата и время создания профиля
        updated_at:
          type: string
          format: date-time
          description: Дата и время последнего обновления профиля
      required:
        - id
        - account_id
        - nickname
        - visibility
        - created_at
        - updated_at
    ProfileUpdate:
      type: object
      properties:
        nickname:
          type: string
          description: Отображаемое имя пользователя
        bio:
          type: string
          description: Биография пользователя
        country:
          type: string
          description: Страна пользователя
        city:
          type: string
          description: Город пользователя
        birth_date:
          type: string
          format: date
          description: Дата рождения пользователя
        gender:
          type: string
          description: Пол пользователя
        visibility:
          type: string
          enum: [public, friends, private]
          description: Видимость профиля
    ProfileList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Profile'
        total:
          type: integer
          description: Общее количество профилей
        page:
          type: integer
          description: Текущая страница
        limit:
          type: integer
          description: Количество элементов на странице
      required:
        - items
        - total
        - page
        - limit
    ProfileHistory:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Уникальный идентификатор записи истории
        profile_id:
          type: string
          format: uuid
          description: Идентификатор профиля
        account_id:
          type: string
          format: uuid
          description: Идентификатор аккаунта
        change_type:
          type: string
          enum: [create, update, delete]
          description: Тип изменения
        field_name:
          type: string
          description: Имя измененного поля
        old_value:
          type: string
          description: Предыдущее значение
        new_value:
          type: string
          description: Новое значение
        changed_by:
          type: string
          format: uuid
          description: Идентификатор аккаунта, внесшего изменение
        changed_at:
          type: string
          format: date-time
          description: Дата и время изменения
      required:
        - id
        - profile_id
        - account_id
        - change_type
        - field_name
        - changed_by
        - changed_at
    ProfileHistoryList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProfileHistory'
        total:
          type: integer
          description: Общее количество записей истории
        page:
          type: integer
          description: Текущая страница
        limit:
          type: integer
          description: Количество элементов на странице
      required:
        - items
        - total
        - page
        - limit
...
    Error:
      type: object
      properties:
        code:
          type: string
          description: Код ошибки
        message:
          type: string
          description: Сообщение об ошибке
        details:
          type: object
          description: Дополнительные детали ошибки
      required:
        - code
        - message
  parameters:
...
    ProfileIdParam:
      name: id
      in: path
      required: true
      description: ID профиля
      schema:
        type: string
        format: uuid
...
    PageParam:
      name: page
      in: query
      description: Номер страницы (начиная с 0)
      schema:
        type: integer
        default: 0
        minimum: 0
    LimitParam:
      name: limit
      in: query
      description: Количество элементов на странице
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
  responses:
    BadRequest:
      description: Некорректный запрос
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Не авторизован
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Доступ запрещен
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError:
      description: Внутренняя ошибка сервера
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key